minor:修改节点管理新建任务前端表单渲染方式
