+++ pipeline_plugins/components/collections/sites/open/nodeman/create_task/v2_0.py
@@ -48,8 +48,6 @@ REMOVE_JOB = ["REMOVE_AGENT", "REMOVE_PROXY"]
 # 主机其它参数
 HOST_EXTRA_PARAMS = ["outer_ip", "login_ip", "data_ip"]
 
-BK_NODEMAN_SUPPORT_TJJ = os.environ.get("BK_NODEMAN_SUPPORT_TJJ", "False") == "True"
-
 
 class NodemanCreateTaskService(NodeManBaseService):
     def execute(self, data, parent_data):
@@ -61,10 +59,14 @@ class NodemanCreateTaskService(NodeManBaseService):
         bk_cloud_id = nodeman_op_target.get("nodeman_bk_cloud_id", "")
         node_type = nodeman_op_target.get("nodeman_node_type", "")
 
-        nodeman_tjj_ticket = ""
-        if BK_NODEMAN_SUPPORT_TJJ and hasattr(data.inputs, "nodeman_ticket_combine"):
-            nodeman_ticket_combine = data.inputs.nodeman_ticket_combine
-            nodeman_tjj_ticket = nodeman_ticket_combine.get("nodeman_tjj_ticket", "")
+        nodeman_ticket = data.get_one_of_inputs("nodeman_ticket", {})
+        nodeman_tjj_ticket = nodeman_ticket.get("nodeman_tjj_ticket", "")
+        if nodeman_tjj_ticket:
+            try:
+                nodeman_tjj_ticket = rsa_decrypt_password(nodeman_tjj_ticket, settings.RSA_PRIV_KEY)
+            except Exception:
+                # password is not encrypted
+                pass
 
         nodeman_op_info = data.inputs.nodeman_op_info
         op_type = nodeman_op_info.get("nodeman_op_type", "")
@@ -254,6 +256,4 @@ class NodemanCreateTaskComponent(Component):
     name = _("新建任务")
     code = "nodeman_create_task"
     bound_service = NodemanCreateTaskService
-    form = "%scomponents/atoms/nodeman/create_task/%s" % (
-        settings.STATIC_URL, "v2_1.js" if BK_NODEMAN_SUPPORT_TJJ else "v2_0.js")
-    version = VERSION
+    form = "%scomponents/atoms/nodeman/create_task/v2_0.js" % settings.STATIC_URL
